Changed the reference category for the thresholds. The first category (observed score zero) is now the reference category, and thresholds are estimated for the remaining categories.
Changed the model parameterization. Scores were treated as (m-x) in the model but are now treated as x. This change does not affect the interaction parameters. It does naturally fit with the change in the reference category (see above).

Changed the prior distribution on the thresholds. Some functions used a normal density, others the beta-prime density (on the exponentiated threshold parameters). Now, the beta-prime distribution is used in all functions.

Changed the way that the likelihood, gradient, and hessian are computed. Now, we use pseudolikelihood as the basis and add the information that comes from the prior distribution to that.

Fixed a bug in the optimization code for the Cauchy prior distribution stipulated on the interaction parameters. The second derivative of the Cauchy prior distribution was incorrect.